requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.1
psycopg2-binary==2.9.6